refactor(proxy,http): move reality_proxy middleware to mockforge-proxy (#555 phase 8)
Issue #555 acceptance criteria includes:
> mockforge-proxy would own the reality-proxy + upstream-forwarding logic

The reality-slider middleware (#222) was the last piece of upstream-
forwarding code still living in mockforge-http. Its only foreign dep is
mockforge_core::consistency::UnifiedState, already in the proxy crate's
dep graph.

Changes:
- mockforge-proxy gains rand + thiserror deps for the per-request
  coin-flip and proxy-side error enum, respectively
- mockforge-proxy/src/reality.rs (new) — verbatim move of the
  middleware + RealityProxyConfig + 5 unit tests
- mockforge-http/src/reality_proxy.rs — now a 15-line re-export shim
  so `mockforge_http::reality_proxy::{RealityProxyConfig,
  reality_proxy_middleware}` callers (notably the layer wired up in
  mockforge_http::lib) keep resolving unchanged

Net: -267 LOC from mockforge-http, +288 LOC to mockforge-proxy.

Verified:
- cargo build -p mockforge-proxy -p mockforge-http
- cargo clippy -p mockforge-proxy -p mockforge-http --all-targets -- -D warnings
- cargo test -p mockforge-proxy (23/23 pass, includes 5 reality tests)
- cargo test -p mockforge-http (601/601 pass)
- cargo fmt --all --check
