+++ crates/mockforge-proxy/Cargo.toml
@@ -28,3 +28,8 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 tracing = { workspace = true }
 url = { workspace = true }
+# `reality.rs` (moved from `mockforge-http` under #555 phase 8) needs
+# `rand::random` for the per-request coin-flip and `thiserror` for the
+# proxy-side error enum.
+rand = { workspace = true }
+thiserror = { workspace = true }
